No Particle Crash
No Particle Crash bug is solved. No more crashes if particle is not
found.
I have added isParticle boolean which is false if no particle is found.
If no particle is found, findKeyPoints breaks and is Particle is set to
false.  Also, filterKeyPoints(); displayResults(); functions are skipped
if no particle is found.